+++ fakernaija/providers/names.py
@@ -3,6 +3,7 @@
 import json
 import random
 from pathlib import Path
+from typing import Any
 
 
 class NameProvider:
@@ -14,31 +15,72 @@ class NameProvider:
         Sets the path to the directory containing name data files.
         """
         self.data_path = Path(__file__).parent / "data" / "names"
-        self.first_names = self.load_json(self.data_path / "first_names.json")
-        self.last_names = self.load_json(self.data_path / "last_names.json")
+        self.first_names = self.load_json(
+            self.data_path / "first_names.json",
+            ["tribe", "gender", "name"],
+        )
+        self.last_names = self.load_json(
+            self.data_path / "last_names.json",
+            ["tribe", "name"],
+        )
 
-    def load_json(self, file_path: str | Path) -> list[dict[str, str]]:
-        """Load data from a JSON file.
+    def load_json(
+        self,
+        file_path: str | Path,
+        required_keys: list[str],
+    ) -> list[dict[str, str]]:
+        """Load data from a JSON file and validate its structure.
 
         Args:
-            file_path (Path): The path to the JSON file.
+            file_path (str | Path): The path to the JSON file.
+            required_keys (list[str]): The keys that each entry in the JSON data must have.
 
         Returns:
             list[dict[str, str]]: The data loaded from the JSON file.
 
         Raises:
             FileNotFoundError: If the JSON file is not found.
-            ValueError: If the JSON data is invalid.
+            ValueError: If the JSON data is invalid or missing required keys.
         """
         try:
             with Path(file_path).open(encoding="utf-8") as file:
-                return json.load(file)
+                data = json.load(file)
+                self.validate_json_structure(data, required_keys)
+                return data
         except FileNotFoundError:
             msg = f"File not found: {file_path}"
             raise FileNotFoundError(msg) from None
         except json.JSONDecodeError as exc:
             msg = f"Error decoding JSON from file: {file_path}"
             raise ValueError(msg) from exc
+
+    @staticmethod
+    def validate_json_structure(
+        data: list[dict[str, Any]],
+        required_keys: list[str],
+    ) -> None:
+        """Validate the structure of the JSON data.
+
+        Args:
+            data (list[dict[str, Any]]): The JSON data to validate.
+            required_keys (list[str]): The keys that each entry in the JSON data must have.
+
+        Raises:
+            ValueError: If any entry is missing a required key or contains extra keys.
+        """
+        for entry in data:
+            entry_keys = set(entry.keys())
+            required_keys_set = set(required_keys)
+
+            missing_keys = required_keys_set - entry_keys
+            extra_keys = entry_keys - required_keys_set
+
+            if missing_keys:
+                msg = f"Missing keys {missing_keys} in entry: {entry}"
+                raise ValueError(msg)
+            if extra_keys:
+                msg = f"Invalid keys {extra_keys} in entry: {entry}"
+                raise ValueError(msg)
 
     def get_first_names(
         self,